joy-proposals - extract potentially reusable code into joy-utils
