+++ packages/joy-proposals/src/Proposal/Body.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Card, Header, Button, Icon, Message } from "semantic-ui-react";
-import { ProposalType } from "../runtime/transport";
+import { ProposalType } from "@polkadot/joy-utils/types/proposals";
 import { blake2AsHex } from '@polkadot/util-crypto';
 import styled from 'styled-components';
 import AddressMini from '@polkadot/react-components/AddressMiniJoy';
@@ -8,12 +8,11 @@ import TxButton from '@polkadot/joy-utils/TxButton';
 import { ProposalId } from "@joystream/types/proposals";
 import { MemberId } from "@joystream/types/members";
 import ProfilePreview from "@polkadot/joy-utils/MemberProfilePreview";
-import { useTransport } from "../runtime";
-import { usePromise } from "../utils";
+import { useTransport, usePromise } from "@polkadot/joy-utils/react/hooks";
 import { Profile } from "@joystream/types/members";
 import { Option } from "@polkadot/types/";
 import { formatBalance } from "@polkadot/util";
-import PromiseComponent from "./PromiseComponent";
+import { PromiseComponent } from "@polkadot/joy-utils/react/components";
 
 type BodyProps = {
   title: string;
@@ -45,7 +44,7 @@ function ProposedMember(props: { memberId?: MemberId | number | null }) {
 
   const transport = useTransport();
   const [ member, error, loading ] = usePromise<Option<Profile> | null>(
-    () => transport.memberProfile(memberId),
+    () => transport.members.memberProfile(memberId),
     null
   );
