+++ packages/joy-proposals/src/Proposal/ChooseProposalType.tsx
@@ -2,10 +2,8 @@ import React, { useState } from "react";
 import ProposalTypePreview from "./ProposalTypePreview";
 import { Item, Dropdown } from "semantic-ui-react";
 
-import { useTransport } from "../runtime";
-import { usePromise } from "../utils";
-import Error from "./Error";
-import Loading from "./Loading";
+import { useTransport, usePromise } from "@polkadot/joy-utils/react/hooks";
+import { PromiseComponent } from "@polkadot/joy-utils/react/components";
 import "./ChooseProposalType.css";
 import { RouteComponentProps } from "react-router-dom";
 
@@ -22,35 +20,31 @@ export type Category = typeof Categories[keyof typeof Categories];
 export default function ChooseProposalType(props: RouteComponentProps) {
   const transport = useTransport();
 
-  const [proposalTypes, error, loading] = usePromise(() => transport.proposalsTypesParameters(), []);
+  const [proposalTypes, error, loading] = usePromise(() => transport.proposals.proposalsTypesParameters(), []);
   const [category, setCategory] = useState("");
-
-  if (loading && !error) {
-    return <Loading text="Fetching proposals..." />;
-  } else if (error || proposalTypes == null) {
-    return <Error error={error} />;
-  }
 
   console.log({ proposalTypes, loading, error });
   return (
     <div className="ChooseProposalType">
-      <div className="filters">
-        <Dropdown
-          placeholder="Category"
-          options={Object.values(Categories).map(category => ({ value: category, text: category }))}
-          value={category}
-          onChange={(e, data) => setCategory((data.value || "").toString())}
-          clearable
-          selection
-        />
-      </div>
-      <Item.Group>
-        {proposalTypes
-          .filter(typeInfo => !category || typeInfo.category === category)
-          .map((typeInfo, idx) => (
-            <ProposalTypePreview key={`${typeInfo} - ${idx}`} typeInfo={typeInfo} history={props.history} />
-          ))}
-      </Item.Group>
+      <PromiseComponent error={error} loading={loading} message={'Fetching proposals\' parameters...'}>
+        <div className="filters">
+          <Dropdown
+            placeholder="Category"
+            options={Object.values(Categories).map(category => ({ value: category, text: category }))}
+            value={category}
+            onChange={(e, data) => setCategory((data.value || "").toString())}
+            clearable
+            selection
+          />
+        </div>
+        <Item.Group>
+          {proposalTypes
+            .filter(typeInfo => !category || typeInfo.category === category)
+            .map((typeInfo, idx) => (
+              <ProposalTypePreview key={`${typeInfo} - ${idx}`} typeInfo={typeInfo} history={props.history} />
+            ))}
+        </Item.Group>
+      </PromiseComponent>
     </div>
   );
 }
