+++ packages/joy-proposals/src/Proposal/ProposalTypePreview.tsx
@@ -4,8 +4,8 @@ import { History } from "history";
 import { Item, Icon, Button, Label } from "semantic-ui-react";
 
 import { Category } from "./ChooseProposalType";
-import { ProposalType } from "../runtime";
-import { slugify, splitOnUpperCase } from "../utils";
+import { ProposalType } from "@polkadot/joy-utils/types/proposals";
+import { slugify, splitOnUpperCase } from "@polkadot/joy-utils/functions/misc";
 import styled from 'styled-components';
 import useVoteStyles from './useVoteStyles';
 import { formatBalance } from "@polkadot/util";
