+++ packages/joy-proposals/src/Proposal/Votes.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Header, Divider, Table, Icon } from "semantic-ui-react";
 import useVoteStyles from "./useVoteStyles";
-import { ProposalVote } from "../runtime";
+import { ProposalVote } from "@polkadot/joy-utils/types/proposals";
 import { VoteKind } from "@joystream/types/proposals";
 import { VoteKindStr } from "./VotingSection";
 import ProfilePreview from "@polkadot/joy-utils/MemberProfilePreview";
