+++ packages/joy-proposals/src/forms/SetContentWorkingGroupMintCapForm.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import { default as MintCapacityForm } from './MintCapacityForm';
 import { RouteComponentProps } from 'react-router';
-import { useTransport } from "../runtime";
-import { usePromise } from "../utils";
-import PromiseComponent from '../Proposal/PromiseComponent';
+import { useTransport, usePromise } from "@polkadot/joy-utils/react/hooks";
+import { PromiseComponent } from "@polkadot/joy-utils/react/components";
 
 const ContentWorkingGroupMintCapForm = (props: RouteComponentProps) => {
   const transport = useTransport();
-  const [ mintCapacity, error, loading ] = usePromise<number>(() => transport.WGMintCap(), 0);
+  const [ mintCapacity, error, loading ] = usePromise<number>(() => transport.contentWorkingGroup.currentMintCap(), 0);
 
   return (
     <PromiseComponent error={error} loading={loading} message="Fetching current mint capacity...">
