--- packages/joy-proposals/src/runtime/index.ts
@@ -1,4 +1,0 @@
-export { ParsedProposal, ProposalType, ProposalVote, IStorageRoleParameters, StorageRoleParameters } from "./transport";
-export { SubstrateTransport } from "./transport.substrate";
-export { MockTransport } from "./transport.mock";
-export { SubstrateProvider, useTransport } from "./TransportContext";
