--- packages/joy-proposals/src/runtime/transport.mock.ts
@@ -1,16 +1,0 @@
-import { Transport, ParsedProposal } from "./transport";
-
-function delay(ms: number) {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-export class MockTransport extends Transport {
-  constructor() {
-    super();
-  }
-
-  async proposals() {
-    await delay(Math.random() * 2000);
-    return Promise.all((Array.from({ length: 5 }, (_, i) => "Not implemented") as unknown) as ParsedProposal[]);
-  }
-}
