+++ packages/joy-utils/src/consts/members.ts
@@ -1,0 +1,3 @@
+import BN from 'bn.js';
+
+export const FIRST_MEMBER_ID = new BN(0);
