+++ packages/joy-utils/src/react/components/index.tsx
@@ -1,0 +1,1 @@
+export { default as PromiseComponent } from "./PromiseComponent";
