+++ packages/joy-utils/src/react/hooks/usePromise.tsx
@@ -1,0 +1,26 @@
+import { useState, useEffect, useCallback } from "react";
+
+export default function usePromise<T>(promise: () => Promise<T>, defaultValue: T): [T, any, boolean, () => Promise<void|null>] {
+  const [state, setState] = useState<{
+    value: T;
+    error: any;
+    isPending: boolean;
+  }>({ value: defaultValue, error: null, isPending: true });
+
+  let isSubscribed = true;
+  const execute = useCallback(() => {
+    return promise()
+      .then(value => (isSubscribed ? setState({ value, error: null, isPending: false }) : null))
+      .catch(error => (isSubscribed ? setState({ value: defaultValue, error: error, isPending: false }) : null));
+  }, [promise]);
+
+  useEffect(() => {
+    execute();
+    return () => {
+      isSubscribed = false;
+    };
+  }, []);
+
+  const { value, error, isPending } = state;
+  return [value, error, isPending, execute];
+}
