+++ packages/joy-utils/src/react/hooks/useTransport.tsx
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+import Transport from "../../transport";
+import { TransportContext } from "../context";
+
+export default function useTransport() {
+  return useContext(TransportContext) as Transport;
+}
